moved handelr functions to different file Removed old doc line
